Time: 1974 ms (24.79%), Space: 40.5 MB (50.43%) - LeetHub

+++ 1230-toss-strange-coins/1230-toss-strange-coins.py
@@ -1,5 +1,5 @@
 class Solution:
-    def findProbability(self, index: int, target: int, n: int, prob: List[float], memo: List[List[float]]) -> int:
+    def findProbability(self, index: int, target: int, n: int, prob: List[float], memo: List[List[float]]) -> float:
         if target < 0:
             return 0
